Fix incorrect type that causes crash when half points are awarded

+++ internal/models/f1/race_result.py
@@ -10,7 +10,7 @@ class RaceResult(BaseModel):
     number: int
     position: int
     positionText: str
-    points: int
+    points: float
     Driver: Driver
     Constructor: Constructor
     grid: int
